WinGui: Removing the dependency on Newtonsoft.Json library in HandBrake.Interop and HandBrake.worker.  Using System.Text.Json instead.    UI will be updated seperately.

+++ win/CS/HandBrake.Interop/Interop/Json/Encode/Video.cs
@@ -70,7 +70,7 @@ namespace HandBrake.Interop.Interop.Json.Encode
         /// <summary>
         /// Gets or sets the quality.
         /// </summary>
-        public double? Quality { get; set; }
+        public decimal? Quality { get; set; }
 
         /// <summary>
         /// Gets or sets the tune.
